+++ src/sd_card_driver.h
@@ -14,7 +14,7 @@ extern "C" {
 
 int init_sd();
 void save_to_sd(char* filename, uint8_t* bytes, int bytesToWrite);
-void write_as_csv_buffered(uint64_t time, float wind, int hx711);
+void write_as_csv_buffered(uint64_t time, float wind, float hx711);
 void write_bytes_buffered(const uint8_t* bytes, int bytesToWrite);
 void write_remaining_buffer();
 int get_number_of_filename(char* filename);
